Review @sk1p
* Test case for visualize_frame
* Split in several test files
* Docstrings
* Imports

Closes #10
